move address helpers into server action scope, remove lib, other minor updates
