+++ client/actions/form.js
@@ -2,13 +2,6 @@
 
 import { redirect } from "next/navigation";
 import { serverUrl } from "@/constants";
-import {
-  formatName,
-  formatAddress,
-  formatDateCreated,
-  formatPhone,
-  formatTime,
-} from "@/lib/fetchAppointments";
 
 export async function requestAppt(formValues) {
   const response = await fetch(serverUrl + "form", {
@@ -33,7 +26,42 @@ export async function fetchAppointments() {
     throw new Error("Failed to fetch appointments");
   }
   const data = await response.json();
-  console.log("🚀 ~ fetchAppointments ~ data:", data);
+
+  // Address filtering helper functions
+
+  const formatName = (name) => {
+    const [firstName, ...rest] = name.split(" ");
+    const lastName = rest.join(" ");
+    if (!lastName) {
+      return firstName;
+    }
+    return `${lastName}, ${firstName}`;
+  };
+
+  const formatTime = (hour) => {
+    return hour <= 12 ? `${hour}a` : `${hour - 12}p`;
+  };
+
+  const formatDateCreated = (dateString) => {
+    const date = new Date(dateString);
+    return `${date.toLocaleDateString()} ${date.toLocaleTimeString([], {
+      hour: "2-digit",
+      minute: "2-digit",
+    })}`;
+  };
+
+  const formatPhone = (phone) => {
+    const cleaned = ("" + phone).replace(/\D/g, "");
+    const match = cleaned.match(/^(\d{3})(\d{3})(\d{4})$/);
+    return match ? `(${match[1]}) ${match[2]}-${match[3]}` : phone;
+  };
+
+  const formatAddress = (address) => {
+    const [firstPart, ...rest] = address.split(",");
+    const secondPart = rest.join(",").trim();
+    return `${firstPart}\n${secondPart}`;
+  };
+
   return data.map((item, index) => ({
     id: item.id || index + 1,
     name: formatName(item.name),
