+++ client/app/admin-dashboard/page.jsx
@@ -1,12 +1,12 @@
 import { Suspense } from "react";
-import { Typography } from "@mui/material";
+import { CircularProgress, Typography } from "@mui/material";
 import ReservationTable from "../../components/admin_dashboard/ReservationTable";
 
 export default function AdminDashboardView() {
   return (
     <div>
       <Typography variant="h1">Reservations:</Typography>
-      <Suspense fallback={<p>Loading...</p>}>
+      <Suspense fallback={<CircularProgress />}>
         <ReservationTable />
       </Suspense>
     </div>
