+++ server/src/routes/appointments.routes.js
@@ -1,5 +1,4 @@
 import { Router } from "express";
-import mongoose from "mongoose";
 import { getAllAppointments } from "../controllers/appointments.controller.js";
 
 const router = Router();
